--- README.md
@@ -1,13 +1,0 @@
-# /* SQLi Vulnerable Scanner */
-
-This will give you the SQLi Vulnerable Website Just by Adding the Dork.
-Feel free to report me bugs.
-This tool have 3 main .pys [pygoogle, pydorker, sqliscan.py]
-To use this tool, run sqliscan.py
-
-This works only on *nix.
-
-Credit given to Xero who wrote pydorker.py
-# Hades.y2k
-
-![alt tag](http://i.imgur.com/jKo4JPZ.png)
