+++ readme.md
@@ -1,0 +1,7 @@
+Author  - Hades.y2k
+License - GNU v2.0
+#Run sqlivul.py to use this tool
+Credit  - Xero (who wrote pydorker.py)
+
+#You can report me for bugs.
+https://github.com/Hadesy2k/sqlivulscan
